home/requirements.txt
certifi
chardet
defusedxml
Django==2.2.10
django-allauth
django-crispy-forms
django-debug-toolbar
django-environ
django-invitations
idna
numpy==1.16.2
oauthlib
pandas
python-dateutil
python3-openid
pytz
PyYAML
requests
requests-oauthlib
six
sqlparse
urllib3
huey
